backend-ocr/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.3
colorama==0.4.6
fastapi==0.136.1
h11==0.16.0
idna==3.13
numpy==2.4.4
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.1
pdf2image==1.17.0
pillow==12.2.0
pydantic==2.13.3
pydantic_core==2.46.3
pytesseract==0.3.13
python-multipart==0.0.26
pyzbar==0.1.9
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
